removed zod todo schemas and added unified them with other schemas
