+++ .gitignore
@@ -4,3 +4,4 @@ prisma/dev.db
 .vscode
 prisma/dev.db-journal
 public
+*.js
